Split local registry module
